requirements/dev.txt
apturl==0.5.2
asn1crypto==0.24.0
Brlapi==0.6.6
cairocffi==0.8.1
CairoSVG==2.1.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
command-not-found==0.3
cryptography==2.2.2
cssselect2==0.2.1
cupshelpers==1.0
defer==1.0.6
defusedxml==0.5.0
distro-info==0.18
Django==2.0.4
django-allauth==0.35.0
django-colorfield==0.1.14
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-dotenv==1.4.2
django-easy-pdf==0.1.1
django-formset-js-improved==0.5.0.2
django-jquery-js==3.1.1
django-mail-templated==2.6.5
django-nexmo==2.0.0a1
django-render-block==0.5
django-templated-email==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
html5lib==1.0.1
httplib2==0.9.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
Mako==1.0.7
MarkupSafe==1.0
mate-hud==18.4.8
mate-tweak==18.4.16
nexmo==2.0.0
oauth==1.0.1
oauthlib==2.0.7
olefile==0.45.1
onboard==1.4.1
pbr==4.0.3
pdfrw==0.4
pexpect==4.2.1
Pillow==5.1.0
psutil==5.4.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycairo==1.16.2
pycparser==2.18
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.6.1
PyPDF2==1.26.0
Pyphen==0.9.4
python-apt==1.6.0
python-dateutil==2.7.2
python-debian==0.1.32
python-xapp==1.0.0
python-xlib==0.20
python3-openid==3.1.0
pytz==2018.4
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
sessioninstaller==0.0.0
setproctitle==1.1.10
simplejson==3.13.2
six==1.11.0
systemd-python==234
testresources==2.0.1
tinycss2==0.6.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
WeasyPrint==0.42.3
webencodings==0.5.1
xhtml2pdf==0.2.2
xkit==0.0.0
zope.interface==4.3.2
